# src/api/swagger/swaggerJson.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the Swagger Documentation implementation of API's on NodeJS. You will find detailed documentation and model data about the REST API's. To check if an API is working , execute the API after entering the valid parameters",
		"version": "1.0.0",
		"title": "Boiler",
		"termsOfService": "",
		"contact": {
			"email": "api@boiler.com"
		}
	},
	"host": "127.0.0.1:4001",
	"basePath": "/api/v1",
...
	"schemes": [
		"http"
	],
	"paths": {
...
		"/admin/deal-details/{dealId}": {
			"get": {
				"tags": ["Admin"],
				"summary": "deal details",
				"description": "get details of a deal",
				"operationId": "dealDetails",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"description": "accessToken",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "dealId",
						"description": "Id of the deal",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Successfully getting deal details"
					},
					"409": {
						"description": "Invalid input Error"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
		"/seller/property": {
			"post": {
				"tags": [
					"SELLER"
				],
				"summary": "Seller adds a property",
				"description": "Seller adds a property",
				"operationId": "postProperty",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "Seller adds a property",
						"required": true,
						"schema": {
							"$ref": "#/definitions/postProperty"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/seller/property/{id}": {
			"get": {
				"tags": [
					"SELLER"
				],
				"summary": " Fetch Specific Property",
				"description": "Fetch Specific Property",
				"operationId": "getProperty",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "uuid",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"delete": {
				"tags": [
					"SELLER"
				],
				"summary": " Soft delete specific property",
				"description": "Soft delete specific property",
				"operationId": "deleteProperty",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "uuid",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"patch": {
				"tags": [
					"SELLER"
				],
				"summary": " ico details by amdin",
				"description": "ico details list by amdin based on icoid",
				"operationId": "getIcoDetails via icoID",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "uuid",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Seller updates a property",
						"required": true,
						"schema": {
							"$ref": "#/definitions/updateProperty"
						}
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/seller/properties/{page}/{limit}": {
			"get": {
				"tags": [
					"SELLER"
				],
				"summary": "Fetch All Properties",
				"description": "Fetch All Properties",
				"operationId": "getProperties",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "page",
						"description": "page number",
						"required": true
					},
					{
						"in": "path",
						"name": "limit",
						"description": "number of properties to show on page",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "success"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"postProperty": {
			"type": "object",
			"required": [
				"city",
				"locality",
				"address",
				"numberOfBathrooms",
				"numberOfBedrooms",
				"numberOfRooms",
				"livingArea",
				"email"
			],
			"properties": {
				"propertyName": {
					"type": "string",
					"example": "2 BHK Dubai Marina Penthouse Villa"
				},
				"subCategory": {
					"type": "string",
					"example": "Villa"
				},
				"showAddressOnWebsite": {
					"type": "boolean",
					"example": true
				},
				"sellerPrice": {
					"type": "number",
					"example": "50000"
				},
				"agentPrice": {
					"type": "number",
					"example": "40000"
				},
				"floorNumber": {
					"type": "number",
					"example": "14"
				},
				"totalFloors": {
					"type": "number",
					"example": "150"
				},
				"city": {
					"type": "string",
					"example": "Dubai"
				},
				"locality": {
					"type": "string",
					"example": "Burj Al Arab"
				},
				"address": {
					"type": "string",
					"example": "1 Sheikh Mohammed bin Rashid Blvd - Dubai - United Arab Emirates"
				},
				"numberOfBathrooms": {
					"type": "number",
					"example": "15"
				},
				"numberOfBedrooms": {
					"type": "number",
					"example": "15"
				},
				"numberOfRooms": {
					"type": "number",
					"example": "15"
				},
				"livingArea": {
					"type": "number",
					"example": "2500"
				},
				"email": {
					"type": "string",
					"example": "john@snow.com"
				}
			},
			"xml": {
				"name": "postProperty"
			}
		},
		"updateProperty": {
			"type": "object",
			"properties": {
				"city": {
					"type": "string",
					"example": "ABC"
				},
				"locality": {
					"type": "string",
					"example": "DEF"
				},
				"address": {
					"type": "string",
					"example": "GHI"
				},
				"numberOfBathrooms": {
					"type": "number",
					"example": "1"
				},
				"numberOfBedrooms": {
					"type": "number",
					"example": "1"
				},
				"numberOfRooms": {
					"type": "number",
					"example": "1"
				},
				"livingArea": {
					"type": "number",
					"example": "2500"
				},
				"email": {
					"type": "string",
					"example": "john@snow.com"
				}
			},
			"xml": {
				"name": "updateProperty"
			}
		}
	}
}